requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.3.0
build==1.1.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
envtoml==0.1.2
filelock==3.13.1
flake8==7.0.0
identify==2.5.35
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pillow==10.3.0
pip-tools==7.4.1
platformdirs==4.2.0
pre-commit==3.6.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rich==13.7.1
rpds-py==0.18.0
setuptools==70.0.0
sqlparse==0.5.0
toml==0.10.2
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
wheel==0.42.0
